Revert original imports in the moved files
