+++ libs/pos/src/test/java/com/woocommerce/android/ui/woopos/util/WooPosConnectionTypeProviderTest.kt
